+++ src/main/java/coupon/coupon/domain/Category.java
@@ -1,4 +1,4 @@
-package coupon.coupon;
+package coupon.coupon.domain;
 
 public enum Category {
