+++ src/main/java/coupon/coupon/domain/Coupon.java
@@ -1,4 +1,4 @@
-package coupon.coupon;
+package coupon.coupon.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
