+++ src/main/java/coupon/coupon/persistence/CouponJpaRepository.java
@@ -1,5 +1,6 @@
-package coupon.coupon;
+package coupon.coupon.persistence;
 
+import coupon.coupon.domain.Coupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponJpaRepository extends JpaRepository<Coupon, Long> {
