+++ src/main/java/coupon/coupon/service/CouponNotFoundException.java
@@ -1,4 +1,4 @@
-package coupon.coupon;
+package coupon.coupon.service;
 
 public class CouponNotFoundException extends RuntimeException {
     public CouponNotFoundException(long couponId) {
