+++ src/main/java/coupon/coupon/service/CouponRepository.java
@@ -1,5 +1,6 @@
-package coupon.coupon;
+package coupon.coupon.service;
 
+import coupon.coupon.domain.Coupon;
 import java.util.Optional;
 
 public interface CouponRepository {
