+++ src/main/java/coupon/coupon/service/CouponService.java
@@ -1,5 +1,6 @@
-package coupon.coupon;
+package coupon.coupon.service;
 
+import coupon.coupon.domain.Coupon;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
