+++ src/main/java/coupon/member/service/MemberRepository.java
@@ -1,5 +1,6 @@
-package coupon.member;
+package coupon.member.service;
 
+import coupon.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
