+++ src/main/java/coupon/membercoupon/domain/MemberCoupon.java
@@ -1,4 +1,4 @@
-package coupon.membercoupon;
+package coupon.membercoupon.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
