+++ src/main/java/coupon/membercoupon/service/MemberCouponRepository.java
@@ -1,5 +1,6 @@
-package coupon.membercoupon;
+package coupon.membercoupon.service;
 
+import coupon.membercoupon.domain.MemberCoupon;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
