+++ src/test/java/coupon/coupon/CouponServiceTest.java
@@ -2,6 +2,9 @@ package coupon.coupon;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import coupon.coupon.domain.Category;
+import coupon.coupon.domain.Coupon;
+import coupon.coupon.service.CouponService;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
