+++ src/test/java/coupon/member/MemberTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import coupon.member.domain.Member;
 import org.junit.jupiter.api.Test;
 
 class MemberTest {
