+++ src/test/java/coupon/membercoupon/MemberCouponServiceTest.java
@@ -5,13 +5,17 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import coupon.coupon.Category;
-import coupon.coupon.Coupon;
-import coupon.coupon.CouponIssuanceDatePassedException;
-import coupon.coupon.CouponNotFoundException;
-import coupon.coupon.CouponRepository;
-import coupon.member.Member;
-import coupon.member.MemberRepository;
+import coupon.coupon.domain.Category;
+import coupon.coupon.domain.Coupon;
+import coupon.coupon.service.CouponIssuanceDatePassedException;
+import coupon.coupon.service.CouponNotFoundException;
+import coupon.coupon.service.CouponRepository;
+import coupon.member.domain.Member;
+import coupon.member.service.MemberRepository;
+import coupon.membercoupon.domain.MemberCoupon;
+import coupon.membercoupon.service.MemberCouponIssuedOverException;
+import coupon.membercoupon.service.MemberCouponResponse;
+import coupon.membercoupon.service.MemberCouponService;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
