+++ src/test/java/coupon/membercoupon/MemberCouponTest.java
@@ -3,6 +3,7 @@ package coupon.membercoupon;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import coupon.membercoupon.domain.MemberCoupon;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import org.junit.jupiter.api.Test;
